+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/SearchHighlighter.java
@@ -52,7 +52,9 @@ public final class SearchHighlighter {
 	private static final ResourceBundle MESSAGES = ResourceBundle.getBundle(SearchHighlighter.class.getName());
 
 	private final JTextComponent textComponent;
-	private final Value<String> searchStringValue = Value.nonNull("").build();
+	private final Value<String> searchStringValue = Value.nonNull("")
+					.listener(this::searchAndHighlightResults)
+					.build();
 	private final State caseSensitiveState = State.state();
 	private final Highlighter highlighter = new DefaultHighlighter();
 	private final List<MatchPosition> searchTextPositions = new ArrayList<>();
@@ -214,7 +216,6 @@ public final class SearchHighlighter {
 	}
 
 	private void bindEvents(JTextComponent textComponent) {
-		searchStringValue.addListener(this::searchAndHighlightResults);
 		caseSensitiveState.addListener(this::searchAndHighlightResults);
 		textComponent.getDocument().addDocumentListener((DocumentAdapter) e -> searchAndHighlightResults());
 		selectedSearchTextPosition.addConsumer(selectedSearchPosition -> {