+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.17.41-SNAPSHOT
 ### is.codion.common.core
 - Value.clear() added, PropertyValue.clear() renamed remove() in order to not clash with method from supertype.
+- Value.Builder, methods for adding listeners and consumers added.
 ### is.codion.swing.common.ui
 - InputDialogBuilder.inputValid() and inputValidator() renamed valid() and validator() respectively.
 ### is.codion.swing.fraework.ui